+++ stroom-integrationtest/src/integrationTest/java/stroom/task/server/TestTaskManagerImpl.java
@@ -14,11 +14,12 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executor;
+import java.util.concurrent.RejectedExecutionException;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.IntStream;
 
 public class TestTaskManagerImpl extends AbstractCoreIntegrationTest {
-
     private static final Logger LOGGER = LoggerFactory.getLogger(TestTaskManagerImpl.class);
 
     @Resource
@@ -35,10 +36,8 @@ public class TestTaskManagerImpl extends AbstractCoreIntegrationTest {
 
         final Executor executor = executorProvider.getExecutor(threadPool);
 
-
         CompletableFuture.runAsync(() ->
                 LOGGER.info("Warming up thread pool")).get();
-
 
         AtomicInteger counter = new AtomicInteger();
         final Queue<Thread> threadsUsed = new ConcurrentLinkedQueue<>();
@@ -48,7 +47,7 @@ public class TestTaskManagerImpl extends AbstractCoreIntegrationTest {
                 .mapToObj(i ->
                         CompletableFuture.runAsync(() -> {
                                     ThreadUtil.sleep(50);
-                                    LOGGER.info("Running task %s", i);
+                                    LOGGER.info("Running task {}}", i);
                                     counter.incrementAndGet();
                                     threadsUsed.add(Thread.currentThread());
                                 },
@@ -63,7 +62,7 @@ public class TestTaskManagerImpl extends AbstractCoreIntegrationTest {
                 .distinct()
                 .count();
 
-        LOGGER.info("Threads used: %s", distinctThreads);
+        LOGGER.info("Threads used: {}}", distinctThreads);
 
         Assert.assertTrue(distinctThreads <= poolSize);
 
@@ -90,7 +89,7 @@ public class TestTaskManagerImpl extends AbstractCoreIntegrationTest {
                 .mapToObj(i ->
                         CompletableFuture.runAsync(() -> {
                                     ThreadUtil.sleep(50);
-                                    LOGGER.info("Running task %s", i);
+                                    LOGGER.info("Running task {}", i);
                                     counter.incrementAndGet();
                                     threadsUsed.add(Thread.currentThread());
                                 },
@@ -105,10 +104,163 @@ public class TestTaskManagerImpl extends AbstractCoreIntegrationTest {
                 .distinct()
                 .count();
 
-        LOGGER.info("Threads used: %s", distinctThreads);
+        LOGGER.info("Threads used: {}", distinctThreads);
 
         Assert.assertTrue(distinctThreads <= poolSize);
         LOGGER.info("Finished");
     }
 
+    @Test
+    public void testCompletedExceptionally() {
+        testCompletedExceptionally(executorProvider, false);
+    }
+
+    private void testCompletedExceptionally(final ExecutorProvider executorProvider, final boolean nested) {
+        final AtomicBoolean terminated = new AtomicBoolean();
+        final AtomicBoolean completedNormally = new AtomicBoolean();
+        final AtomicBoolean completedExceptionally = new AtomicBoolean();
+        RuntimeException exception = null;
+
+        try {
+            final Executor executor = executorProvider.getExecutor();
+            CompletableFuture.runAsync(() -> {
+                if (CurrentTaskState.isTerminated()) {
+                    terminated.set(true);
+                }
+                throw new RuntimeException("Expected");
+            }, executor)
+                    .thenRun(() -> completedNormally.set(true))
+                    .exceptionally(t -> {
+                        completedExceptionally.set(true);
+                        return null;
+                    })
+                    .join();
+        } catch (final RuntimeException t) {
+            exception = t;
+            LOGGER.error(t.getMessage(), t);
+            completedExceptionally.set(true);
+        }
+
+        Assert.assertFalse(completedNormally.get());
+        Assert.assertTrue(completedExceptionally.get());
+
+        if (terminated.get()) {
+            throw new TaskTerminatedException();
+        }
+
+        if (nested && exception != null) {
+            throw exception;
+        }
+    }
+
+    @Test
+    public void testCompletedNormally() {
+        final AtomicBoolean completedNormally = new AtomicBoolean();
+        final AtomicBoolean completedExceptionally = new AtomicBoolean();
+
+        final Executor executor = executorProvider.getExecutor();
+        CompletableFuture.runAsync(() -> {
+        }, executor)
+                .thenRun(() -> completedNormally.set(true))
+                .exceptionally(t -> {
+                    completedExceptionally.set(true);
+                    return null;
+                })
+                .join();
+
+        Assert.assertTrue(completedNormally.get());
+        Assert.assertFalse(completedExceptionally.get());
+    }
+
+    @Test
+    public void testCompletedExceptionallyNested() {
+        final AtomicBoolean completedNormally = new AtomicBoolean();
+        final AtomicBoolean completedExceptionally = new AtomicBoolean();
+
+        final Executor executor = executorProvider.getExecutor();
+        CompletableFuture.runAsync(() -> {
+            CurrentTaskState.currentTask().terminate();
+            testCompletedExceptionally();
+        }, executor)
+                .thenRun(() -> completedNormally.set(true))
+                .exceptionally(t -> {
+                    completedExceptionally.set(true);
+                    return null;
+                })
+                .join();
+
+        Assert.assertFalse(completedNormally.get());
+        Assert.assertTrue(completedExceptionally.get());
+    }
+
+    private void testCompletedExceptionallyNested(final ExecutorProvider executorProviderOuter, final ExecutorProvider executorProviderInner) {
+        final AtomicBoolean completedNormally = new AtomicBoolean();
+        final AtomicBoolean completedExceptionally = new AtomicBoolean();
+
+        final Executor executor = executorProviderOuter.getExecutor();
+        CompletableFuture.runAsync(() -> {
+            CurrentTaskState.currentTask().terminate();
+            testCompletedExceptionally(executorProviderInner, true);
+        }, executor)
+                .thenRun(() -> completedNormally.set(true))
+                .exceptionally(t -> {
+                    completedExceptionally.set(true);
+                    return null;
+                })
+                .join();
+
+        Assert.assertFalse(completedNormally.get());
+        Assert.assertTrue(completedExceptionally.get());
+    }
+
+    @Test
+    public void testCompletedNormallyNested() {
+        final AtomicBoolean completedNormally = new AtomicBoolean();
+        final AtomicBoolean completedExceptionally = new AtomicBoolean();
+
+        final Executor executor = executorProvider.getExecutor();
+        CompletableFuture.runAsync(() -> {
+            CurrentTaskState.currentTask().terminate();
+            testCompletedNormally();
+        }, executor)
+                .thenRun(() -> completedNormally.set(true))
+                .exceptionally(t -> {
+                    completedExceptionally.set(true);
+                    return null;
+                })
+                .join();
+
+        Assert.assertTrue(completedNormally.get());
+        Assert.assertFalse(completedExceptionally.get());
+    }
+
+    @Test
+    public void testRejectedExecution() {
+        testCompletedExceptionally(createRejectedExecutorProvider(), false);
+    }
+
+    @Test
+    public void testRejectedNestedExecution() {
+        testCompletedExceptionallyNested(executorProvider, createRejectedExecutorProvider());
+    }
+
+    private ExecutorProvider createRejectedExecutorProvider() {
+        final Executor executor = new Executor() {
+            @Override
+            public void execute(final Runnable command) {
+                throw new RejectedExecutionException("Expected");
+            }
+        };
+        return new ExecutorProvider() {
+            @Override
+            public Executor getExecutor() {
+                return executor;
+            }
+
+            @Override
+            public Executor getExecutor(final ThreadPool threadPool) {
+                return executor;
+            }
+        };
+    }
 }